Testing:  - moved data init logic to management for general access  - added test for all GET urls
